requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
pymysql
werkzeug
pyserial
pyyaml